updog: move migration searching code to update_metadata for reuse
