+++ sources/updater/update_metadata/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 
 [dependencies]
 chrono = { version = "0.4.9", features = ["serde"] }
+lazy_static = "1.2"
 parse-datetime = { path = "../../parse-datetime" }
 rand = "0.7.0"
 regex = "1.1"
@@ -16,7 +17,7 @@ serde = { version = "1.0.100", features = ["derive"] }
 serde_json = "1.0.40"
 serde_plain = "0.3.0"
 snafu = "0.6.0"
-migrator = { path = "../../api/migration/migrator" }
+tough = "0.4.0"
 
 [lib]
 name = "update_metadata"
